Directional Lightmap Trials begin
created a new bu (backup) folder to place the ship files that have both directional lights on and off.

the intent is to run the game only with lightschemes that do have directional lights on, to see how they look and get some research data on their operation.

Also minor tweaks to some lightschemes.